Add Card class for payment source
Why is this change neccesary?

We need to separate the distnitc payment sources, there are cards and
offline recurrent reference

How does it address the issue?

Creating a Card class to respresent the card payment sources.

What side effects does this change have?

Posibly breaking changes to the people who update the library