Add created_at property to Customer class.
Why is this change neccesary?

Its a useful property that we are not including in the class so the
users starts to requested it to us.

How does it address the issue?

Just add the property and the corresponding test.

What side effects does this change have?

Hopely users happiness increses.